fix dtypes.Decimal bug
